refactor: LottoNumber를 Lotto 클래스와 별도의 클래스 파일로 분리
